refactor : Extract utility functions refactor : Reduce complexity through removing needless if
