+++ src/main/kotlin/calcaulator/util/Strings.kt
@@ -1,0 +1,5 @@
+package calcaulator.util
+
+fun String.toIntList(delimiter: Regex = Regex(",")) = this.split(delimiter)
+    .filter { it.isNotBlank() }
+    .map { it.trim().toInt() }
